NAV-25011: Flytter BrevmottakerFeltnavn inn i Form-komponenten for å reflektere det som er gjort andre steder
