+++ src/frontend/Komponenter/Behandling/Brevmottakere/baks/modal/form/felt/Adresselinje1Felt.tsx
@@ -1,9 +1,8 @@
 import { useController, useFormContext } from 'react-hook-form';
 import React from 'react';
 import { Alert, TextField } from '@navikt/ds-react';
-import { BrevmottakerFeltnavn } from './felttyper';
 import { EøsLandkode } from '../../../../../../../Felles/Landvelger/landkode';
-import { BrevmottakerFormValues } from '../BrevmottakerForm';
+import { BrevmottakerFeltnavn, BrevmottakerFormValues } from '../BrevmottakerForm';
 
 interface Props {
     erLesevisning?: boolean;
