+++ src/frontend/Komponenter/Behandling/Brevmottakere/baks/modal/form/felt/Adresselinje2Felt.tsx
@@ -1,8 +1,7 @@
 import { useController, useFormContext } from 'react-hook-form';
 import React from 'react';
 import { TextField } from '@navikt/ds-react';
-import { BrevmottakerFeltnavn } from './felttyper';
-import { BrevmottakerFormValues } from '../BrevmottakerForm';
+import { BrevmottakerFeltnavn, BrevmottakerFormValues } from '../BrevmottakerForm';
 
 interface Props {
     erLesevisning?: boolean;
