+++ src/frontend/Komponenter/Behandling/Brevmottakere/baks/modal/form/felt/PoststedFelt.tsx
@@ -2,8 +2,7 @@ import { useController, useFormContext } from 'react-hook-form';
 import React from 'react';
 import { TextField } from '@navikt/ds-react';
 import { EøsLandkode } from '../../../../../../../Felles/Landvelger/landkode';
-import { BrevmottakerFeltnavn } from './felttyper';
-import { BrevmottakerFormValues } from '../BrevmottakerForm';
+import { BrevmottakerFeltnavn, BrevmottakerFormValues } from '../BrevmottakerForm';
 
 interface Props {
     erLesevisning?: boolean;
