--- src/frontend/Komponenter/Behandling/Brevmottakere/baks/modal/form/felt/felttyper.ts
@@ -1,9 +1,0 @@
-export enum BrevmottakerFeltnavn {
-    MOTTAKERROLLE = 'mottakerRolle',
-    LANDKODE = 'landkode',
-    NAVN = 'navn',
-    ADRESSELINJE1 = 'adresselinje1',
-    ADRESSELINJE2 = 'adresselinje2',
-    POSTNUMMER = 'postnummer',
-    POSTSTED = 'poststed',
-}
